:recycle: Refator: SearchBox 컴포넌트 분리
